// src/generated/type-graphql/resolvers/inputs/UserPurchaseScalarWhereInput.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../../scalars";
import { DateTimeFilter } from "../inputs/DateTimeFilter";
import { EnumOrderStatusFilter } from "../inputs/EnumOrderStatusFilter";
import { IntFilter } from "../inputs/IntFilter";
import { StringFilter } from "../inputs/StringFilter";
import { StringNullableFilter } from "../inputs/StringNullableFilter";

@TypeGraphQL.InputType("UserPurchaseScalarWhereInput", {})
export class UserPurchaseScalarWhereInput {
  @TypeGraphQL.Field(_type => [UserPurchaseScalarWhereInput], {
    nullable: true
  })
  AND?: UserPurchaseScalarWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [UserPurchaseScalarWhereInput], {
    nullable: true
  })
  OR?: UserPurchaseScalarWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [UserPurchaseScalarWhereInput], {
    nullable: true
  })
  NOT?: UserPurchaseScalarWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  id?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  qty?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  size?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => EnumOrderStatusFilter, {
    nullable: true
  })
  status?: EnumOrderStatusFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  razorpayOrderId?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  razorpayPaymentId?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  userId?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  merchId?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  campaignId?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  createdAt?: DateTimeFilter | undefined;
}